Quickfix: Set documentId in Page service
This is just a quickfix. Currently this is a very simple way to get the
application working by passing the documentId through the Page service
into our navigation panel. The navigation panel should not be embedded
into the index.html, because it can only appear with a document - where
the id could be passed on a better way.