Inject EntityManager into services to manage its lifecycle
EntityManagers must be closed after ues, otherwise the database connections
are not reclaimed and the db server dies very quickly.

There is a @PersistenceContext which could be used, but it is not supported
in jersey out of the box, and it needs the persistence unit to be specified
explicitly, which does not fit into our system of different units in tests
than in production.