Add main class for starting the application in Java SE
In glassfish, there is a blocker issue concerning xerces. Gate depends on
xerces being there with the default package names, but glassfish ships with
one in the com.sun.* package. As long as we can not override the glassfish
xerces, we will not be able to ship in glassfish.

Besides of that, it is sometimes easier to debug a standalone application,
and it starts/deploys way faster.