Fix #66 xerces hell.
StanfordNLP and GATE have a dependency to xerces, and at least GATE
directly accesses the org.apache.xerces package.

We can't provide xerces with the endorsed standards override mechanism
because that causes exceptions in glassfish and jboss, probably due to osgi
compilcating the whole process, see [1]. But the maven xerces package
expects to override the endorsed standard using the services directory,
causing a mix of different xml implementations to be used.

This removes the services directory so that xerces is just a plain library,
having nothing to do with JAXP.

[1] https://stackoverflow.com/questions/26852835/use-endorsed-standard-override-mechanism-for-xerces-in-glassfish-4-0