Fix hk2 not closing EntityManagers
The dispose() method is not called. Instead, we can use the
CloseableService class. But that requires that Model is created via
dependency injection, so the *Application classes need to bind it via
class name instead of instances. That in turn means that the constructors
for unit tests models need to be converted into own classes.

Binding the non-dropping model and using the dropping model in tests did
not work any longer, so it has been refactored:  Now, there is one
EntityManagerFactory per test run - stored in a static attribute of
UnitTestModel. It needs to be manually reset. But that allows us to
explicitly drop the data base before/after each test run.