Create sub-resources via main cdi framework instead of hk2
HK2 seens to mangage its own set of beans, so the Model class will be
re-instanciated with every request, and with it the EntityManagerFactory.
That in turn causes the whole database structure to be updated at every
request - which is very slow.

Creating the sub-resources with @Inject will use WELD on glassfish, and
hk2 in the tests and standalone version. hk2 does not discover services
automatically, so they are collected via the @MangagedBean annotation.