Workaround for Chapter.range rest interface
The interface specifies that chapter.range should have a progress property,
which only exists in the Occurrence class. That can only be created when
the document length is present, because this is the only way to calculate
the progress. Currently, Chapter does not know of the document it lies in.

This is temporary workaround until the Chapter class is refactored.