Fix relation occurrences by simplifying them
The previous version counted the number of pair occurrences, which did not
work for more than two entities. Now, the fingerprint of a set of entities
is simply the intersection of all the single fingerprints. This is a bit
different than originally intended, because the "two entities occur next
to each other" now depends on the fingerprint's resolution instead of a
fixed with like in EntityRelationModule. But it is probably what the user
expects, and it is very simple to implement with JPA. Intersection of
regions is not something JPA supports natively, and I do not know if it is
possible at all.