Implement exact results for RelationOccurrencesService
This is not really possible with SQL/JPA, so we do it in Java. As it is
only requested for parts of the document, the performance penality is not
too bad.

This is currently not completely correct: We do not really want the
intersection of the entity occurrences, because that might just be the
middle between them two. E.g. "[Frodo {meets] Bilbo}" - in this case,
the intersection is only "meets". But it works well enough for a first
demo.